feat: integrate Google Fonts and URL launcher, add new screens for sign-in and message display
- Added dependencies for Google Fonts, Timeago, and URL Launcher in pubspec.yaml.
- Introduced new assets including a Google sign-in icon.
- Created SignInCard and MessageCard components for user interaction and message display.
- Refactored main.dart to implement a new app structure with StealthNoteApp and StealthHomePage.
- Updated UI to enhance user experience with a modern design.
