+++ flutter/pubspec.yaml
@@ -28,6 +28,9 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.6
+  google_fonts: ^6.1.0
+  timeago: ^3.5.0
+  url_launcher: ^6.2.1
 
 dev_dependencies:
   integration_test:
@@ -58,6 +61,7 @@ flutter:
     - assets/srs.local
     - assets/jwt_input.json
     - assets/jwt-srs.local
+    - assets/google.png
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware
